feat: move daemon to new binary
