+++ Cargo.toml
@@ -18,6 +18,7 @@ repository = "https://github.com/kantord/enwiro"
 
 [workspace.dependencies]
 enwiro-sdk = { path = "enwiro-sdk", version = "0.2.2" }
+enwiro-daemon = { path = "enwiro-daemon", version = "0.0.1" }
 home = "0.5.9"
 tracing = "0.1"
 tracing-appender = "0.2"
