+++ enwiro-daemon/Cargo.toml
@@ -6,11 +6,18 @@ description = "Background daemon for enwiro: refreshes the recipe cache and forw
 license = "GPL-3.0-or-later"
 repository.workspace = true
 
+[[bin]]
+name = "enwiro-daemon"
+path = "src/main.rs"
+
 [dependencies]
 anyhow = "1"
+confy = "2.0.0"
 dirs = "6"
 enwiro-sdk.workspace = true
+home.workspace = true
 libc = "0.2"
+serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1"
 signal-hook = "0.3"
 system-idle-time = "1.0.4"
