+++ enwiro/src/usage_stats.rs
@@ -1,26 +1,11 @@
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
-use std::time::{SystemTime, UNIX_EPOCH};
 use std::{fs, io};
 
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct UserIntentSignals {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub activation_buffer: Vec<(i64, f64)>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub switch_buffer: Vec<(i64, f64)>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct EnvStats {
-    #[serde(flatten, default)]
-    pub signals: UserIntentSignals,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub cookbook: Option<String>,
-}
+pub use enwiro_daemon::meta::{
+    EnvStats, load_env_meta, now_timestamp, record_activation_per_env, record_cook_metadata_per_env,
+};
 
 /// Per-environment usage statistics.
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
@@ -30,13 +15,6 @@ pub struct UsageStats {
 
 pub fn stats_path() -> Option<PathBuf> {
     dirs::data_dir().map(|d| d.join("enwiro").join("usage-stats.json"))
-}
-
-pub fn now_timestamp() -> i64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_secs() as i64)
-        .unwrap_or(0)
 }
 
 /// Load stats from disk. Returns empty stats on any error (missing file, corrupt JSON, etc.).
@@ -77,90 +55,6 @@ fn record_activation_to(path: &Path, env_name: &str) {
     entry.signals.activation_buffer.truncate(10);
     if let Err(e) = save_stats(path, &stats) {
         tracing::warn!(error = %e, "Could not save usage stats");
-    }
-}
-
-/// Load per-environment metadata from its meta.json file.
-/// Returns default (empty) metadata on any error.
-pub fn load_env_meta(env_dir: &Path) -> EnvStats {
-    let meta_path = env_dir.join("meta.json");
-    fs::read_to_string(&meta_path)
-        .ok()
-        .and_then(|s| serde_json::from_str(&s).ok())
-        .unwrap_or_default()
-}
-
-fn save_env_meta(env_dir: &Path, meta: &EnvStats) -> io::Result<()> {
-    let meta_path = env_dir.join("meta.json");
-    enwiro_sdk::fs::atomic_write(&meta_path, serde_json::to_string(meta)?.as_bytes())
-}
-
-/// Record activation in per-env meta.json. Best-effort.
-pub fn record_activation_per_env(env_dir: &Path) {
-    if !env_dir.is_dir() {
-        return;
-    }
-    let mut meta = load_env_meta(env_dir);
-    meta.signals.activation_buffer.push((now_timestamp(), 1.0));
-    meta.signals
-        .activation_buffer
-        .sort_by_key(|b| std::cmp::Reverse(b.0));
-    meta.signals.activation_buffer.truncate(10);
-    if let Err(e) = save_env_meta(env_dir, &meta) {
-        tracing::warn!(error = %e, "Could not save environment metadata");
-    }
-}
-
-/// Record a workspace switch event in per-env meta.json. Best-effort.
-pub fn record_switch_per_env(env_dir: &Path, timestamp: i64) {
-    if !env_dir.is_dir() {
-        return;
-    }
-    let mut meta = load_env_meta(env_dir);
-
-    // Compute the most recent signal timestamp across both buffers, before the current push.
-    let last_activation_ts = meta
-        .signals
-        .activation_buffer
-        .iter()
-        .map(|&(ts, _)| ts)
-        .max();
-    let last_switch_ts = meta.signals.switch_buffer.iter().map(|&(ts, _)| ts).max();
-    let last_signal_ts = [last_activation_ts, last_switch_ts]
-        .into_iter()
-        .flatten()
-        .max();
-
-    meta.signals.switch_buffer.push((timestamp, 1.0));
-    meta.signals
-        .switch_buffer
-        .sort_by_key(|b| std::cmp::Reverse(b.0));
-    meta.signals.switch_buffer.truncate(25);
-
-    // If the gap since the last signal exceeds 8 hours, inject a synthetic activation.
-    const EIGHT_HOURS: i64 = 28800;
-    if last_signal_ts.is_none_or(|last| timestamp - last > EIGHT_HOURS) {
-        meta.signals.activation_buffer.push((timestamp, 0.4));
-        meta.signals
-            .activation_buffer
-            .sort_by_key(|b| std::cmp::Reverse(b.0));
-        meta.signals.activation_buffer.truncate(10);
-    }
-
-    if let Err(e) = save_env_meta(env_dir, &meta) {
-        tracing::warn!(error = %e, "Could not save switch event");
-    }
-}
-
-/// Save cookbook and description metadata in per-env meta.json. Best-effort.
-pub fn record_cook_metadata_per_env(env_dir: &Path, cookbook: &str, description: Option<&str>) {
-    let mut meta = load_env_meta(env_dir);
-    meta.cookbook = Some(cookbook.to_string());
-    if let Some(d) = description {
-        meta.description = Some(d.to_string());
-    }
-    if let Err(e) = save_env_meta(env_dir, &meta) {
-        tracing::warn!(error = %e, "Could not save environment metadata");
     }
 }
 
@@ -288,6 +182,7 @@ pub fn slot_scores(all_stats: &HashMap<String, EnvStats>, now: i64) -> HashMap<S
 #[cfg(test)]
 mod tests {
     use super::*;
+    use enwiro_daemon::meta::UserIntentSignals;
 
     #[test]
     fn test_record_and_load() {
@@ -330,68 +225,6 @@ mod tests {
     }
 
     #[test]
-    fn test_per_env_record_activation() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        record_activation_per_env(&env_dir);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(meta.signals.activation_buffer.len(), 1);
-        assert!(meta.signals.activation_buffer[0].0 > 0);
-    }
-
-    #[test]
-    fn test_per_env_record_activation_increments() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        record_activation_per_env(&env_dir);
-        record_activation_per_env(&env_dir);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(meta.signals.activation_buffer.len(), 2);
-    }
-
-    #[test]
-    fn test_per_env_record_cook_metadata() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        record_cook_metadata_per_env(&env_dir, "github", Some("Fix auth bug"));
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(meta.cookbook, Some("github".to_string()));
-        assert_eq!(meta.description, Some("Fix auth bug".to_string()));
-    }
-
-    #[test]
-    fn test_per_env_load_missing_dir_returns_default() {
-        let meta = load_env_meta(Path::new("/nonexistent/env/dir"));
-        assert!(meta.signals.activation_buffer.is_empty());
-        assert_eq!(meta.description, None);
-    }
-
-    #[test]
-    fn test_per_env_metadata_preserves_activation_on_cook() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        record_activation_per_env(&env_dir);
-        record_activation_per_env(&env_dir);
-        record_cook_metadata_per_env(&env_dir, "git", Some("My project"));
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(meta.signals.activation_buffer.len(), 2);
-        assert_eq!(meta.cookbook, Some("git".to_string()));
-        assert_eq!(meta.description, Some("My project".to_string()));
-    }
-
-    #[test]
     fn test_load_missing_file_returns_empty() {
         let stats = load_stats(Path::new("/nonexistent/path/stats.json"));
         assert!(stats.envs.is_empty());
@@ -405,129 +238,6 @@ mod tests {
 
         let stats = load_stats(&path);
         assert!(stats.envs.is_empty());
-    }
-
-    // ── activation_buffer ring-buffer tests ────────────────────────────────
-
-    /// `UserIntentSignals` must have an `activation_buffer` field of type `Vec<(i64, f64)>`.
-    /// A freshly default-constructed value must have an empty buffer.
-    #[test]
-    fn test_activation_buffer_field_exists_and_defaults_empty() {
-        let signals = UserIntentSignals::default();
-        assert!(
-            signals.activation_buffer.is_empty(),
-            "activation_buffer must be empty by default"
-        );
-    }
-
-    /// `activation_buffer` must be serialised under that exact JSON key name,
-    /// flattened into the top-level `EnvStats` object.
-    #[test]
-    fn test_env_stats_serializes_activation_buffer_field_name() {
-        let signals = UserIntentSignals {
-            activation_buffer: vec![(1_700_000_000i64, 1.0f64)],
-            ..Default::default()
-        };
-        let stats = EnvStats {
-            signals,
-            ..Default::default()
-        };
-        let json = serde_json::to_string(&stats).unwrap();
-        let value: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert!(
-            value.get("activation_buffer").is_some(),
-            "activation_buffer must appear as a top-level JSON key in EnvStats"
-        );
-    }
-
-    /// Recording an activation pushes a `(timestamp, 1.0)` entry into `activation_buffer`.
-    #[test]
-    fn test_record_activation_pushes_entry_to_buffer() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let before = now_timestamp();
-        record_activation_per_env(&env_dir);
-        let after = now_timestamp();
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            1,
-            "buffer must have exactly one entry after one activation"
-        );
-        let (ts, weight) = meta.signals.activation_buffer[0];
-        assert!(
-            ts >= before && ts <= after,
-            "buffer entry timestamp must be within the recording window"
-        );
-        assert!(
-            (weight - 1.0).abs() < 1e-10,
-            "buffer entry weight must be 1.0"
-        );
-    }
-
-    /// Each subsequent activation must append another `(timestamp, 1.0)` entry.
-    #[test]
-    fn test_record_activation_appends_entries() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        record_activation_per_env(&env_dir);
-        record_activation_per_env(&env_dir);
-        record_activation_per_env(&env_dir);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            3,
-            "buffer must grow by one entry per activation"
-        );
-        for &(_, weight) in &meta.signals.activation_buffer {
-            assert!((weight - 1.0).abs() < 1e-10, "every weight must be 1.0");
-        }
-    }
-
-    /// The buffer must be capped at N=10: when an 11th activation is recorded,
-    /// the oldest entry is dropped so the buffer stays at length 10.
-    #[test]
-    fn test_activation_buffer_capped_at_ten() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        // Pre-populate the buffer with 10 entries using a fixed old timestamp.
-        let old_ts: i64 = 1_000_000;
-        let mut meta = load_env_meta(&env_dir);
-        for _ in 0..10 {
-            meta.signals.activation_buffer.push((old_ts, 1.0));
-        }
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        // Record one more activation — should trim to 10.
-        record_activation_per_env(&env_dir);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            10,
-            "buffer must stay at capacity 10 after an 11th activation"
-        );
-
-        // The newest entry (just recorded) must be present at buffer[0] after descending sort.
-        let newest_ts = meta.signals.activation_buffer[0].0;
-        assert!(
-            newest_ts > old_ts,
-            "newest entry should be present after cap; got ts={newest_ts}"
-        );
-        // The oldest surviving entries must still carry the old timestamp.
-        assert_eq!(
-            meta.signals.activation_buffer.last().unwrap().0,
-            old_ts,
-            "some pre-populated old entries must still be in the buffer after cap"
-        );
     }
 
     /// `frecency_score` must return 0.0 for an empty buffer.
@@ -1048,500 +758,6 @@ mod tests {
              slot={}, launcher={}",
             slot["switch-only"],
             launcher["switch-only"]
-        );
-    }
-
-    // ── switch_buffer / record_switch_per_env tests ────────────────────────
-
-    /// `UserIntentSignals` must have a `switch_buffer` field of type `Vec<(i64, f64)>`.
-    /// A freshly default-constructed value must have an empty buffer.
-    #[test]
-    fn test_switch_buffer_field_exists_and_defaults_empty() {
-        let signals = UserIntentSignals::default();
-        assert!(
-            signals.switch_buffer.is_empty(),
-            "switch_buffer must be empty by default"
-        );
-    }
-
-    /// `switch_buffer` must be serialised under that exact JSON key name,
-    /// flattened into the top-level `EnvStats` object.
-    #[test]
-    fn test_env_stats_serializes_switch_buffer_field_name() {
-        let signals = UserIntentSignals {
-            switch_buffer: vec![(1_700_000_000i64, 1.0f64)],
-            ..Default::default()
-        };
-        let stats = EnvStats {
-            signals,
-            ..Default::default()
-        };
-        let json = serde_json::to_string(&stats).unwrap();
-        let value: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert!(
-            value.get("switch_buffer").is_some(),
-            "switch_buffer must appear as a top-level JSON key in EnvStats"
-        );
-    }
-
-    /// `record_switch_per_env` must push a `(timestamp, 1.0)` entry into `switch_buffer`.
-    #[test]
-    fn test_record_switch_per_env_pushes_entry() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.switch_buffer.len(),
-            1,
-            "switch_buffer must have exactly one entry after one call"
-        );
-        let (stored_ts, weight) = meta.signals.switch_buffer[0];
-        assert_eq!(
-            stored_ts, ts,
-            "stored timestamp must match the provided one"
-        );
-        assert!(
-            (weight - 1.0).abs() < 1e-10,
-            "switch_buffer entry weight must be 1.0"
-        );
-    }
-
-    /// Each subsequent call to `record_switch_per_env` must append another entry.
-    #[test]
-    fn test_record_switch_per_env_appends_entries() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        record_switch_per_env(&env_dir, 1_700_000_001);
-        record_switch_per_env(&env_dir, 1_700_000_002);
-        record_switch_per_env(&env_dir, 1_700_000_003);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.switch_buffer.len(),
-            3,
-            "switch_buffer must grow by one entry per call"
-        );
-        for &(_, weight) in &meta.signals.switch_buffer {
-            assert!((weight - 1.0).abs() < 1e-10, "every weight must be 1.0");
-        }
-    }
-
-    /// The buffer must be kept in descending timestamp order.
-    #[test]
-    fn test_record_switch_per_env_buffer_sorted_descending() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        record_switch_per_env(&env_dir, 1_700_000_001);
-        record_switch_per_env(&env_dir, 1_700_000_003);
-        record_switch_per_env(&env_dir, 1_700_000_002);
-
-        let meta = load_env_meta(&env_dir);
-        let timestamps: Vec<i64> = meta
-            .signals
-            .switch_buffer
-            .iter()
-            .map(|&(ts, _)| ts)
-            .collect();
-        let mut sorted = timestamps.clone();
-        sorted.sort_by(|a, b| b.cmp(a));
-        assert_eq!(
-            timestamps, sorted,
-            "switch_buffer entries must be in descending timestamp order"
-        );
-    }
-
-    /// The buffer must be capped at N=25: when a 26th entry is recorded,
-    /// the oldest entry is dropped so the buffer stays at length 25.
-    #[test]
-    fn test_switch_buffer_capped_at_twenty_five() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        // Pre-populate the buffer with 25 entries using a fixed old timestamp.
-        let old_ts: i64 = 1_000_000;
-        let mut meta = load_env_meta(&env_dir);
-        for _ in 0..25 {
-            meta.signals.switch_buffer.push((old_ts, 1.0));
-        }
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        // Record one more switch — should trim to 25.
-        let new_ts: i64 = 1_700_000_000;
-        record_switch_per_env(&env_dir, new_ts);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.switch_buffer.len(),
-            25,
-            "switch_buffer must stay at capacity 25 after a 26th entry"
-        );
-
-        // The newest entry must be present at switch_buffer[0] after descending sort.
-        let newest_ts = meta.signals.switch_buffer[0].0;
-        assert_eq!(
-            newest_ts, new_ts,
-            "newest entry should be present after cap; got ts={newest_ts}"
-        );
-        // The oldest surviving entries must still carry the old timestamp.
-        assert_eq!(
-            meta.signals.switch_buffer.last().unwrap().0,
-            old_ts,
-            "some pre-populated old entries must still be in the buffer after cap"
-        );
-    }
-
-    /// `record_switch_per_env` must not affect `activation_buffer`.
-    #[test]
-    fn test_record_switch_per_env_does_not_touch_activation_buffer() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        record_activation_per_env(&env_dir);
-        record_switch_per_env(&env_dir, 1_700_000_000);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            1,
-            "activation_buffer must be untouched by record_switch_per_env"
-        );
-    }
-
-    /// Old on-disk JSON without `switch_buffer` must deserialize to an empty switch_buffer.
-    #[test]
-    fn test_old_json_without_switch_buffer_gives_empty_buffer() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-        fs::write(
-            env_dir.join("meta.json"),
-            r#"{"activation_buffer":[[1700000000,1.0]]}"#,
-        )
-        .unwrap();
-
-        let meta = load_env_meta(&env_dir);
-        assert!(
-            meta.signals.switch_buffer.is_empty(),
-            "old JSON without switch_buffer key must deserialize to an empty buffer"
-        );
-    }
-
-    /// `record_switch_per_env` must be a no-op when the environment directory does not exist.
-    /// It must NOT create the directory or write any file — switch events for unknown
-    /// environments are silently discarded.
-    #[test]
-    fn test_record_switch_per_env_does_nothing_for_nonexistent_dir() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("nonexistent-env");
-        // Precondition: the directory must not exist before the call.
-        assert!(
-            !env_dir.exists(),
-            "test precondition: env_dir must not exist before calling record_switch_per_env"
-        );
-
-        record_switch_per_env(&env_dir, 1_700_000_000);
-
-        assert!(
-            !env_dir.exists(),
-            "record_switch_per_env must not create the env directory when it does not exist"
-        );
-    }
-
-    // ── synthetic activation on switch tests ──────────────────────────────
-
-    /// When both `activation_buffer` and `switch_buffer` are empty (no prior signals),
-    /// a call to `record_switch_per_env` must inject a synthetic activation entry
-    /// `(timestamp, 0.4)` into `activation_buffer`, because inactivity is unbounded
-    /// (trivially exceeds the 8-hour threshold).
-    #[test]
-    fn test_record_switch_per_env_adds_synthetic_activation_when_no_prior_signals() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            1,
-            "activation_buffer must contain exactly one synthetic entry after the first-ever switch"
-        );
-        let (entry_ts, entry_weight) = meta.signals.activation_buffer[0];
-        assert_eq!(
-            entry_ts, ts,
-            "synthetic activation timestamp must match the switch timestamp"
-        );
-        assert!(
-            (entry_weight - 0.4).abs() < 1e-10,
-            "synthetic activation weight must be 0.4, got {entry_weight}"
-        );
-    }
-
-    /// When the env's last activation is more than 8 hours before `timestamp`,
-    /// a call to `record_switch_per_env` must inject a synthetic activation entry
-    /// `(timestamp, 0.4)` into `activation_buffer`.
-    #[test]
-    fn test_record_switch_per_env_adds_synthetic_activation_when_last_activation_over_8h_ago() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        let eight_hours_plus_one: i64 = 8 * 3600 + 1;
-        let old_activation_ts = ts - eight_hours_plus_one;
-
-        // Pre-populate activation_buffer with one old activation (>8h before ts).
-        let mut meta = load_env_meta(&env_dir);
-        meta.signals
-            .activation_buffer
-            .push((old_activation_ts, 1.0));
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        // activation_buffer must now have 2 entries: the old one + the synthetic one.
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            2,
-            "activation_buffer must gain a synthetic entry when last activation was >8h ago"
-        );
-        // The newest entry (at index 0 after descending sort) must be the synthetic one.
-        let (newest_ts, newest_weight) = meta.signals.activation_buffer[0];
-        assert_eq!(
-            newest_ts, ts,
-            "synthetic activation timestamp must match the switch timestamp"
-        );
-        assert!(
-            (newest_weight - 0.4).abs() < 1e-10,
-            "synthetic activation weight must be 0.4, got {newest_weight}"
-        );
-    }
-
-    /// When the env's last switch is more than 8 hours before `timestamp` and
-    /// `activation_buffer` is empty, `record_switch_per_env` must inject a synthetic
-    /// activation `(timestamp, 0.4)` because the switch buffer's most-recent entry
-    /// determines inactivity.
-    #[test]
-    fn test_record_switch_per_env_adds_synthetic_activation_when_last_switch_over_8h_ago() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        let eight_hours_plus_one: i64 = 8 * 3600 + 1;
-        let old_switch_ts = ts - eight_hours_plus_one;
-
-        // Pre-populate only switch_buffer with one old entry (>8h before ts);
-        // activation_buffer remains empty.
-        let mut meta = load_env_meta(&env_dir);
-        meta.signals.switch_buffer.push((old_switch_ts, 1.0));
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            1,
-            "activation_buffer must contain one synthetic entry when last switch was >8h ago"
-        );
-        let (entry_ts, entry_weight) = meta.signals.activation_buffer[0];
-        assert_eq!(
-            entry_ts, ts,
-            "synthetic activation timestamp must match the switch timestamp"
-        );
-        assert!(
-            (entry_weight - 0.4).abs() < 1e-10,
-            "synthetic activation weight must be 0.4, got {entry_weight}"
-        );
-    }
-
-    /// When the env's most recent signal (activation or switch) is within 8 hours of
-    /// `timestamp`, `record_switch_per_env` must NOT inject a synthetic activation entry.
-    #[test]
-    fn test_record_switch_per_env_no_synthetic_activation_when_last_signal_within_8h() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        let one_hour: i64 = 3600;
-        let recent_activation_ts = ts - one_hour; // 1h before switch → within 8h
-
-        // Pre-populate activation_buffer with a recent activation (within 8h).
-        let mut meta = load_env_meta(&env_dir);
-        meta.signals
-            .activation_buffer
-            .push((recent_activation_ts, 1.0));
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        // activation_buffer must still have exactly 1 entry — the original; no synthetic added.
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            1,
-            "activation_buffer must remain unchanged when last signal was within 8h"
-        );
-        let (entry_ts, entry_weight) = meta.signals.activation_buffer[0];
-        assert_eq!(
-            entry_ts, recent_activation_ts,
-            "the only entry must be the pre-existing activation, not a synthetic one"
-        );
-        assert!(
-            (entry_weight - 1.0).abs() < 1e-10,
-            "pre-existing activation weight must still be 1.0, got {entry_weight}"
-        );
-    }
-
-    // ── synthetic-activation boundary / mixed-buffer tests ────────────────
-
-    /// T1a: when the gap is exactly 8 hours (28800s), the threshold is NOT crossed
-    /// (`>` is strict), so no synthetic activation must be injected.
-    #[test]
-    fn test_record_switch_per_env_no_synthetic_at_exact_8h_boundary() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        const EIGHT_HOURS: i64 = 28800;
-        let boundary_activation_ts = ts - EIGHT_HOURS; // gap == 28800, condition is false
-
-        let mut meta = load_env_meta(&env_dir);
-        meta.signals
-            .activation_buffer
-            .push((boundary_activation_ts, 1.0));
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            1,
-            "activation_buffer must NOT gain a synthetic entry when gap == 28800 (boundary is exclusive)"
-        );
-        assert_eq!(
-            meta.signals.activation_buffer[0].0, boundary_activation_ts,
-            "the only entry must still be the pre-existing one, not a synthetic"
-        );
-    }
-
-    /// T1b: when the gap is exactly 28801s (one second past 8 hours), the threshold IS crossed
-    /// (`timestamp - last > 28800` is true), so a synthetic activation must be injected.
-    #[test]
-    fn test_record_switch_per_env_synthetic_injected_at_28801s_gap() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        let just_over_ts = ts - 28801; // gap == 28801, condition is true
-
-        let mut meta = load_env_meta(&env_dir);
-        meta.signals.activation_buffer.push((just_over_ts, 1.0));
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            2,
-            "activation_buffer must gain a synthetic entry when gap == 28801 (just over threshold)"
-        );
-        let (newest_ts, newest_weight) = meta.signals.activation_buffer[0];
-        assert_eq!(
-            newest_ts, ts,
-            "synthetic activation timestamp must match the switch timestamp"
-        );
-        assert!(
-            (newest_weight - 0.4).abs() < 1e-10,
-            "synthetic activation weight must be 0.4, got {newest_weight}"
-        );
-    }
-
-    /// T2a: `activation_buffer` has an old entry (>8h) but `switch_buffer` has a recent
-    /// entry (<8h). The max of both is recent, so no synthetic activation must be injected.
-    /// This verifies that `last_switch_ts` participates in the max computation.
-    #[test]
-    fn test_record_switch_per_env_no_synthetic_when_switch_buffer_is_recent_activation_is_old() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        let old_activation_ts = ts - (8 * 3600 + 1); // >8h ago
-        let recent_switch_ts = ts - 3600; // 1h ago (within 8h)
-
-        let mut meta = load_env_meta(&env_dir);
-        meta.signals
-            .activation_buffer
-            .push((old_activation_ts, 1.0));
-        meta.signals.switch_buffer.push((recent_switch_ts, 1.0));
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            1,
-            "activation_buffer must NOT gain a synthetic entry when switch_buffer has a recent entry (1h ago)"
-        );
-        assert_eq!(
-            meta.signals.activation_buffer[0].0, old_activation_ts,
-            "the only entry must still be the original old activation, not a synthetic"
-        );
-    }
-
-    /// T2b: `switch_buffer` has an old entry (>8h) but `activation_buffer` has a recent
-    /// entry (<8h). The max of both is recent, so no synthetic activation must be injected.
-    /// This verifies that `last_activation_ts` participates in the max computation.
-    #[test]
-    fn test_record_switch_per_env_no_synthetic_when_activation_buffer_is_recent_switch_is_old() {
-        let dir = tempfile::tempdir().unwrap();
-        let env_dir = dir.path().join("my-project");
-        fs::create_dir(&env_dir).unwrap();
-
-        let ts: i64 = 1_700_000_000;
-        let old_switch_ts = ts - (8 * 3600 + 1); // >8h ago
-        let recent_activation_ts = ts - 3600; // 1h ago (within 8h)
-
-        let mut meta = load_env_meta(&env_dir);
-        meta.signals.switch_buffer.push((old_switch_ts, 1.0));
-        meta.signals
-            .activation_buffer
-            .push((recent_activation_ts, 1.0));
-        save_env_meta(&env_dir, &meta).unwrap();
-
-        record_switch_per_env(&env_dir, ts);
-
-        let meta = load_env_meta(&env_dir);
-        assert_eq!(
-            meta.signals.activation_buffer.len(),
-            1,
-            "activation_buffer must NOT gain a synthetic entry when activation_buffer has a recent entry (1h ago)"
-        );
-        assert_eq!(
-            meta.signals.activation_buffer[0].0, recent_activation_ts,
-            "the only entry must still be the original recent activation, not a synthetic"
         );
     }
