+++ justfile
@@ -15,6 +15,9 @@ install-dev:
         pkill -x enw && sleep 0.2 || true
     fi
     installed=$(cargo install --list | grep -E '^enwiro' | awk '{print $1}')
+    # enwiro-daemon is a core binary; include it even on first-time setups
+    # where it isn't yet in `cargo install --list`.
+    echo "$installed" | grep -qx 'enwiro-daemon' || installed="$installed enwiro-daemon"
     for crate in $installed; do
         [ "$crate" = "enwiro-sdk" ] && continue
         # The `enwiro` crate produces a binary named `enw`; all other crates
@@ -46,6 +49,7 @@ install-release:
     #!/usr/bin/env bash
     set -euo pipefail
     installed=$(cargo install --list | grep -E '^enwiro' | awk '{print $1}')
+    echo "$installed" | grep -qx 'enwiro-daemon' || installed="$installed enwiro-daemon"
     for crate in $installed; do
         [ "$crate" = "enwiro-sdk" ] && continue
         echo "Installing $crate from crates.io..."
